Test if cache file is writable before unlink it

+++ system/libraries/Cache/drivers/Cache_file.php
@@ -276,7 +276,7 @@ class CI_Cache_file extends CI_Driver {
 
 		if ($data['ttl'] > 0 && time() > $data['time'] + $data['ttl'])
 		{
-			unlink($this->_cache_path.$id);
+			file_exists($this->_cache_path.$id) && unlink($this->_cache_path.$id);
 			return FALSE;
 		}
